refactor: 오예림님 코드리뷰 추가 작성
